Feat: 활동지역 추가/조회

+++ src/main/java/com/hihi/square/domain/buyer/repository/BuyerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BuyerRepository extends JpaRepository<Buyer, Long> {
+public interface BuyerRepository extends JpaRepository<Buyer, Integer> {
     Optional<Buyer> findByUidAndMethod(String uid, LoginMethod method);
 }
